Refactor of neurite scaling
1) Add a new way a setting the growth parameters for the apical dendrite growth

If the parameter file has a key ["context_constraints"]["apical"]["target"]
then it is assumed that the new scaling algorithm is used for the apical.
Otherwise the standard way of setting the input parameters is used.

This key is interpreted as the expected max length of the apical and the value
is computed from the original biological cells, a target_layer and position within it (target_layer_position).

If key ["context_constraints"]["apical"]["target"] is found, it is assumed that
the following keys are present:
- ["context_constraints"]["apical"]["target"]["slope"]
- ["context_constraints"]["apical"]["target"]["intercept"]
- ["context_constraints"]["apical"]["target"]["layer"]
- ["context_constraints"]["apical"]["target"]["fraction"]

2) Add a dendritic rescaling post-growth to conform with hard limits constraints

Despite the scaling of the input growth parameters, the grown dendrites can be
sometimes too short or too long. In this case they can exceed the constraints
given by the hard limit rules.
The post-growth rescaling will:
- downscale the dendrites if they exceed the max hard limit rule
- upscale the dendrites if they do not reach the min hard limit rule

Change-Id: I2cdf3a68e58b7912958e0c5cda49010c9f9c01e8