fix issue in max y extend with orientation
previously, we use COLS.Y as the direction for max extent,
but in a non flat atlas, the max extent should corresponds to
the y direction in the atlas, not the y coordinates of the points
in the morphology. We thus project the points to the atlas
orientation vector.

in addition, two global variables are defines to allow external control
on when region-grower will crash with two small target lengths for
growing cells

Change-Id: I86d2062a9bf5be760f3bfb887082847e46da5ddb